requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
coverage==5.1
cryptography==2.8
daphne==2.4.1
dj-database-url==0.5.0
Django==3.0.4
django-redis==4.11.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.3
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
more-itertools==8.2.0
msgpack==0.6.2
msgpack-python==0.5.6
numpy==1.18.2
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.1
pytest-asyncio==0.11.0
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
wcwidth==0.1.9
whitenoise==5.0.1
zipp==3.1.0
zope.interface==5.1.0
